requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
stripe==14.0.1
psycopg2-binary==2.9.10
drf-spectacular==0.29.0
redis==7.1.0
pymongo==4.15.4
